+++ src/interface/buddyjs.js
@@ -3,7 +3,6 @@
 var Detector = require('buddy.js/lib/detector');
 var reporters = require('buddy.js/lib/reporters');
 var grunt = require('grunt');
-var p = require('path');
 
 module.exports = function (options, f) {
   options.args = grunt.file.expand(options.args);
@@ -11,7 +10,7 @@ module.exports = function (options, f) {
   // Retrieve the requested reporter, defaulting to simple
   var ReporterType = reporters[options.reporter] || reporters.simple;
   var detector = new Detector(options.args, options.enforceConst, options.ignore);
-  var reporter = new ReporterType(detector);
+  new ReporterType(detector);
 
   var found = 0;
 
@@ -19,5 +18,6 @@ module.exports = function (options, f) {
     found++;
   });
 
-  detector.run().then(f, f);
+  detector.run().then(f).
+  catch (f);
 };