+++ src/interface/jscs.js
@@ -1,10 +1,11 @@
-var Cli = require('jscs/lib/cli');
-var shjs = require("shelljs");
+'use strict';
+
+var cli = require('jscs/lib/cli');
 var grunt = require('grunt');
 
 module.exports = function (options, f) {
   options = options || {};
   options.args = grunt.file.expand(options.args);
-  Cli(options);
+  cli(options);
   f();
 };