+++ HISTORY.md
@@ -1,4 +1,12 @@
 
+### v1.0.3 - 10/20/2014
+
+- check-build should pass check-build :')
+
+### v1.0.2 - 10/20/2014
+
+- fix default `.checkbuild`
+
 ### v1.0.1 - 10/19/2014
 
 - nsp now uses `package` instead of `shrinkwrap`